@@ -1,4 +1,4 @@
-package gitlab
+package oicd_generic
 
 import (
 	"encoding/json"
@@ -7,9 +7,10 @@ import (
 	"net/http"
 	"net/url"
 	"oauth2-proxy-nexus3/authprovider"
-)
+	"oauth2-proxy-nexus3/logger"
 
-const userInfoEndpointPath = "/oauth/userinfo"
+	"go.uber.org/zap"
+)
 
 // Client implements `authprovider.Client`.
 type Client struct {
@@ -18,35 +19,36 @@ type Client struct {
 
 // GetUserInfo implements `authprovider.Client`.
 func (s *Client) GetUserInfo(accessToken string) (authprovider.UserInfo, error) {
-	endpoint, err := url.Parse(fmt.Sprintf(s.URL.String() + userInfoEndpointPath))
+	endpoint, err := url.Parse(s.URL.String())
 	if err != nil {
-		return nil, fmt.Errorf("Failed to parse the GitLab URL: %s", err)
+		return nil, fmt.Errorf("Failed to parse the OpenID Connect URL: %s", err)
 	}
 
 	req, err := http.NewRequest("GET", endpoint.String(), nil)
 	if err != nil {
-		return nil, fmt.Errorf("failed to create the GitLab GET userinfo request: %s", err)
+		return nil, fmt.Errorf("failed to create the OpenID Connect GET userinfo request: %s", err)
 	}
 
 	req.Header.Set("Authorization", "Bearer "+accessToken)
 
 	res, err := http.DefaultClient.Do(req)
 	if err != nil {
-		return nil, fmt.Errorf("failed to request the GitLab GET userinfo endpoint on %s: %s", s.URL, err)
+		return nil, fmt.Errorf("failed to request the OpenID Connect GET userinfo endpoint on %s: %s", s.URL, err)
 	}
 	defer res.Body.Close()
 
+	logger.Debug("oidc_generic provider response", zap.Any("response", res.Body), zap.Any("header", res.Header))
 	if res.StatusCode != http.StatusOK {
 		if resBody, err := io.ReadAll(res.Body); err == nil {
-			return nil, fmt.Errorf("failed to request the GitLab GET userinfo endpoint on %s: %s", s.URL, resBody)
+			return nil, fmt.Errorf("failed to request the OpenID Connect GET userinfo endpoint on %s: %s", s.URL, resBody)
 		}
 
-		return nil, fmt.Errorf("failed to read the GitLab GET userinfo error response: %s", err)
+		return nil, fmt.Errorf("failed to read the OpenID Connect GET userinfo error response: %s", err)
 	}
 
 	var userInfo UserInfo
 	if err := json.NewDecoder(res.Body).Decode(&userInfo); err != nil {
-		return nil, fmt.Errorf("failed to decode the GitLab GET userinfo responses: %s", err)
+		return nil, fmt.Errorf("failed to decode the OpenID Connect GET userinfo responses: %s", err)
 	}
 
 	return &userInfo, nil
@@ -54,11 +56,9 @@ func (s *Client) GetUserInfo(accessToken string) (authprovider.UserInfo, error) 
 
 // UserInfo implements `authprovider.UserInfo`.
 type UserInfo struct {
-	// GitLab does not offer Given and Family names,
-	// https://docs.gitlab.com/ee/integration/openid_connect_provider.html
-	User       string `json:"nickname"`
-	GivenName  string
-	FamilyName string
+	User       string   `json:"nickname"`
+	GivenName  string   `json:"given_name,omitempty"`
+	FamilyName string   `json:"family_name,omitempty"`
 	Email      string   `json:"email"`
 	Groups     []string `json:"groups"`
 }
@@ -68,12 +68,12 @@ func (s *UserInfo) Username() string {
 	return s.User
 }
 
-// GivenName implements `authprovider.UserInfo`.
+// Givenname implements `authprovider.UserInfo`.
 func (s *UserInfo) Givenname() string {
 	return s.GivenName
 }
 
-// FamilyName implements `authprovider.UserInfo`.
+// Familyname implements `authprovider.UserInfo`.
 func (s *UserInfo) Familyname() string {
 	return s.FamilyName
 }
