@@ -5,6 +5,7 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"oauth2-proxy-nexus3/authprovider/gitlab"
+	"oauth2-proxy-nexus3/config"
 	"oauth2-proxy-nexus3/nexus"
 	"testing"
 
@@ -18,15 +19,19 @@ func TestNew(t *testing.T) {
 		oauthAccessToken = "token"
 		nexusUser        = nexus.User{
 			UserID:       "foo",
+			FirstName:    "",
+			LastName:     "",
 			EmailAddress: "foo@test.bar",
 			RoleIDs:      []string{"bar"},
 		}
 		nexusAvailablesRoles = []nexus.Role{{ID: nexusUser.RoleIDs[0]}}
 
 		gitlabOIDCTestSrv = gitlab.NewTestServer(oauthAccessToken, &gitlab.UserInfo{
-			Nickname: nexusUser.UserID,
-			Email:    nexusUser.EmailAddress,
-			Groups:   nexusUser.RoleIDs,
+			User:       nexusUser.UserID,
+			GivenName:  nexusUser.FirstName,
+			FamilyName: nexusUser.LastName,
+			Email:      nexusUser.EmailAddress,
+			Groups:     nexusUser.RoleIDs,
 		})
 		gitlabOIDCTestSrvURL, _ = url.Parse(gitlabOIDCTestSrv.URL)
 
@@ -37,13 +42,19 @@ func TestNew(t *testing.T) {
 		nexusTestSrvURL, _ = url.Parse(nexusTestSrv.URL)
 
 		rproxyAccessTokenHeader = "X-Forwarded-Access-Token"
-		rproxy                  = New(
-			nexusTestSrvURL, gitlabOIDCTestSrvURL, nexusTestSrvURL,
-			rproxyAccessTokenHeader,
-			"null", "null", "X-Forwarded-User",
-		)
 
-		rProxySrv = httptest.NewServer(rproxy.Router.GetRoute(routeName).GetHandler())
+		cfg = config.Config{
+			NexusURL:                      nexusTestSrvURL,
+			AuthProvider:                  "gitlab",
+			AuthProviderURL:               gitlabOIDCTestSrvURL,
+			AuthProviderAccessTokenHeader: rproxyAccessTokenHeader,
+			NexusAdminUser:                "null",
+			NexusAdminPassword:            "null",
+			NexusRutHeader:                "X-Forwarded-User",
+		}
+		rproxy = Run(&cfg)
+
+		rProxySrv = httptest.NewServer(rproxy.Router)
 	)
 
 	defer gitlabOIDCTestSrv.Close()
@@ -59,5 +70,5 @@ func TestNew(t *testing.T) {
 
 	res, err = rProxySrv.Client().Do(sucessfulReq)
 	require.NoError(t, err)
-	require.Equal(t, http.StatusNotFound, res.StatusCode)
+	require.Equal(t, http.StatusBadRequest, res.StatusCode)
 }
